requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain
langchain-community
langchain-google-genai
pdfplumber
pandas
PyMuPDF
python-multipart
Pillow
faiss-cpu
pdf2image
opencv-python
pytesseract
python-docx
openpyxl
unstructured[xlsx]
langchain_experimental

# fix numpy version
numpy<2

# torch CPU
torch==2.2.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html

# embeddings
sentence-transformers
langchain-huggingface
PyJWT


langchain-core>=0.2.0
langgraph>=0.0.40
langchain>=0.2.0
typing-extensions>=4.6.1
mysql-connector-python

redisvl
redis
langchain-redis